requirements.txt
flask
playwright
python-docx
beautifulsoup4
gunicorn
